// src/mediasoup/enums/rtc-events.ts
export enum RTC_EVENTS {
  CREATE_ROOM = 'CREATE_ROOM',
  JOIN_ROOM = 'JOIN_ROOM',
  CREATE_TRANSPORT = 'CREATE_TRANSPORT',
  CONNECT_TRANSPORT = 'CONNECT_TRANSPORT',
  PRODUCE = 'PRODUCE',
  CONSUME = 'CONSUME',
  CREATE_DATA_PRODUCER = 'CREATE_DATA_PRODUCER',
  CREATE_DATA_CONSUMER = 'CREATE_DATA_CONSUMER',
  SEND_DATA = 'SEND_DATA',
}
export const EVENT_NAME = 'rtc-event';